+++ wcli_arginfo.h
@@ -222,6 +222,10 @@ ZEND_END_ARG_INFO()
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_wcli_is_cmd_call, 0, 0, _IS_BOOL, 0)
 ZEND_END_ARG_INFO()
 
+ZEND_BEGIN_ARG_WITH_RETURN_TYPE_MASK_EX(arginfo_wcli_where, 0, 0, MAY_BE_BOOL|MAY_BE_STRING)
+	ZEND_ARG_TYPE_INFO(0, file, IS_STRING, 0)
+ZEND_END_ARG_INFO()
+
 
 // ********************************************************************
 // *************************** DECLARATIONS ***************************
@@ -284,6 +288,7 @@ ZEND_FUNCTION(wcli_set_position);
 ZEND_FUNCTION(wcli_get_module_path);
 ZEND_FUNCTION(wcli_get_parent_pid);
 ZEND_FUNCTION(wcli_is_cmd_call);
+ZEND_FUNCTION(wcli_where);
 
 
 
@@ -350,6 +355,7 @@ static const zend_function_entry ext_functions[] = {
 	ZEND_FE(wcli_get_module_path, arginfo_wcli_get_module_path)
 	ZEND_FE(wcli_get_parent_pid, arginfo_wcli_get_parent_pid)
 	ZEND_FE(wcli_is_cmd_call, arginfo_wcli_is_cmd_call)
+	ZEND_FE(wcli_where, arginfo_wcli_where)